backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
openpyxl==3.1.2
